Client Side Validation is working now
