+++ app/models/user.rb
@@ -5,6 +5,14 @@ class User < ApplicationRecord
          :recoverable, :rememberable, :validatable
 
   # validations
-  validates :first_name, :last_name, :birth_date, :cpf, :phone, presence: true, if: -> (user) { !user.admin }
-  validates :password, format: { with: /([A-Za-z]+[0-9]|[0-9]+[A-Za-z]).+/ }
+  validates :first_name, :last_name, :birth_date, :cpf, :phone, presence: true, unless: :admin?
+  validates :password, format: { with: /(.*([A-Za-z]+[0-9]|[0-9]+[A-Za-z]).*)/ }
+
+  # private
+
+  # def password_format
+  #   return if password.blank? || !password.match?(/([A-Za-z]+[0-9]|[0-9]+[A-Za-z])/)
+
+  #   errors.add(:password, 'com formato inválido')
+  # end
 end
