+++ config/locales/pt-BR.yml
@@ -1,0 +1,13 @@
+pt-BR:
+  activerecord:
+    attributes:
+      user:
+        email: E-mail
+        password: Senha
+        password_confirmation: Confirmação de senha
+        first_name: Nome
+        last_name: Sobrenome
+        birth_date: Data de nascimento
+        cpf: CPF
+        phone: Telefone
+        address: Endereço
